refactor: move handler into deployment package and break it down to multiple files
